feat: components 폴더 생성 및 세일 alert 분리
