--- src/basic/utils/generateText.js
@@ -1,4 +1,0 @@
-
-export const generateText = (target, text) => {
-	target.textContent = text;
-};
